Transformers cleanup
