+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Launch Vite App",
+      "type": "chrome",
+      "request": "launch",
+      "url": "http://localhost:5050",
+      "webRoot": "${workspaceFolder}/apps/viewer",
+      "preLaunchTask": "Build and Start"
+    },
+    {
+      "name": "Run Tests Transform",
+      "type": "node",
+      "request": "launch",
+      "runtimeExecutable": "npx",
+      "runtimeArgs": [
+        "lerna",
+        "run",
+        "test",
+        "--scope",
+        "@allmaps/transform"
+      ],
+      "cwd": "${workspaceFolder}",
+      "console": "integratedTerminal",
+      "internalConsoleOptions": "neverOpen"
+    }
+  ]
+}
