+++ packages/transform/test/input/gcps-test.js
@@ -1,0 +1,81 @@
+export const gcps_3 = [
+  {
+    image: [518, 991],
+    world: [4.9516614, 52.4633102]
+  },
+  {
+    image: [4345, 2357],
+    world: [5.0480391, 52.5123762]
+  },
+  {
+    image: [2647, 475],
+    world: [4.9702906, 52.5035815]
+  }
+]
+
+export const gcps_6 = [
+  {
+    image: [1344, 4098],
+    world: [4.4091165, 51.9017125]
+  },
+  {
+    image: [4440, 3441],
+    world: [4.5029222, 51.9164451]
+  },
+  {
+    image: [3549, 4403],
+    world: [4.4764224, 51.897309]
+  },
+  {
+    image: [1794, 2130],
+    world: [4.4199066, 51.9391509]
+  },
+  {
+    image: [3656, 2558],
+    world: [4.4775683, 51.9324358]
+  },
+  {
+    image: [2656, 3558],
+    world: [4.4572643, 51.9143043]
+  }
+]
+
+export const gcps_10 = [
+  { image: [6933, 3641], world: [-5.6931398, 56.1290282] },
+  {
+    image: [6158, 1470],
+    world: [-6.2921755, 58.5220974]
+  },
+  {
+    image: [9142, 2240],
+    world: [-2.1077545, 57.687078]
+  },
+  {
+    image: [9626, 9058],
+    world: [-1.5642528, 50.6648133]
+  },
+  {
+    image: [8387, 1447],
+    world: [-3.0589806, 58.6152482]
+  },
+  {
+    image: [11709, 8644],
+    world: [1.5829224, 50.8667829]
+  },
+  {
+    image: [8995, 5647],
+    world: [-2.8041724, 54.0553335]
+  },
+  {
+    image: [12903, 6620],
+    world: [4.7685316, 52.9630668]
+  },
+  {
+    image: [8262, 9534],
+    world: [-3.6827528, 50.2116403]
+  },
+  {
+    image: [6925, 9356],
+    world: [-5.7572694, 50.055299]
+  }
+]
